Move Template image fields table and image preview canvas logic to reusable components
